+++ functions/file/global.js
@@ -1,8 +1,13 @@
 const { mdFile } = require('./md.js');
+const { createFile } = require('./create.js');
 
 let global_object = new Object();
 class fileManager {
     constructor() {}
+
+    create() {
+        createFile();
+    }
 
     md() {
         mdFile();