+++ functions/render/global.js
@@ -1,0 +1,17 @@
+const { fileRender } = require('./file.js');
+const { textRender } = require('./text.js');
+
+let global_object = new Object();
+class renderManager {
+    constructor() {}
+
+    file() {
+        fileRender();
+    }
+
+    text() {
+        textRender();
+    }
+}
+
+module.exports = renderManager;